requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain-community
google-generativeai
